Implement August 1st 2017 (UAHF) hardfork

+++ wire/common.go
@@ -601,7 +601,7 @@ func ReadVarString(r io.Reader, pver uint32) (string, error) {
 	// Prevent variable length strings that are larger than the maximum
 	// message size.  It would be possible to cause memory exhaustion and
 	// panics without a sane upper bound on this count.
-	if count > MaxMessagePayload {
+	if count > uint64(MaxMessagePayload) {
 		str := fmt.Sprintf("variable length string is too long "+
 			"[count %d, max %d]", count, MaxMessagePayload)
 		return "", messageError("ReadVarString", str)
